+++ app/p2p/store/index.ts
@@ -11,7 +11,9 @@ export const store = configureStore({
     }
 })
 
+if (typeof window !== 'undefined') {
 window.store = store
+}
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
 export type RootState = ReturnType<typeof store.getState>
